backend/User/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
supabase==2.0.3